backend/requirements.txt
opencv-python
pyzbar 
numpy
pydantic
PyYAML
rich
python-multipart
fastapi
uvicorn[standard]
websockets
pillow
aiofiles
python-dotenv
python-jose
pydantic-settings